pyproject.toml
[project]
name = "levindorge-back"
version = "0.0.9"
requires-python = ">=3.13"
dependencies = [
    "cleanstack",
    "fastapi>=0.116.1",
    "logfire[fastapi]>=4.1.0",
    "polyfactory>=2.22.1",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.13.2",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
cleanstack = { git = "https://github.com/julienpillaud/cleanstack", tag = "0.4.0" }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff==0.12.7",
]

[tool.ruff]
output-format = "concise"
[tool.ruff.lint]
select = [
    "E", "W",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
    "PL",  # Pylint
    "RUF",
]
ignore = [
    "UP046",  # non-pep695-generic-class
    "UP047",  # non-pep695-generic-function
]

[tool.mypy]
strict = true
plugins = ['pydantic.mypy']

[tool.logfire]
ignore_no_config=true

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)30s - %(message)s"
